Install Ruby scripts under $prefix/share/ruby
This has a few advantages over putting them in Ruby's `sitelibdir`:

1. When using a custom prefix (e.g. the way Homebrew does), all of the files
   will end up under that one prefix.
2. We're no longer overloading the semantics of dist_pkgdata_DATA (although,
   arguably we could have switched to using pkgruby_SCRIPTS without changing
   the install target directory)
3. We can't `require 'gitsh/foo'` from other Ruby programs without modifying
   the `$LOAD_PATH` first. Since we're distributing an application, not a
   library, it seemed strange for the internals to be globally available. If
   we wanted that behaviour we should split out the library code into a gem.

The nested `eval "eval ..."` calls are to account for the various $*dir
variables being defined in terms of other variables. The prefix check is
copied from the generated configure file, but needs to be done earlier for
these evals to work. Based on this mailing list post:
http://lists.gnu.org/archive/html/autoconf/2004-09/msg00040.html