Build bin/gitsh with make not configure
This is based on the advice here:
http://www.gnu.org/software/autoconf/manual/autoconf.html#Installation-Directory-Variables

In order to support a prefix set by either configure or make, the prefix
should not be hard coded in a file generated by configure (other than a
Makefile).

Note: This would not be necessary if we went back to installing the Ruby
scripts in Ruby's `sitelibdir`. See 0be4a3f645dd9bedb94d3fd8d5ef230af02b9b44