Vendorize gems for distribution.
This adds a vendor directory, whose Makefile knows how to vendorize all the
gems from the `:dist` bundler group, and write a `setup.rb` file to update
the `$LOAD_PATH` to include each gem.

The `vendor/gems` target is marked `PHONY` to avoid it being included in
`make all` and therefore happening on an end user's machine, however it is a
dependency of `make dist` so the vendorized gems will be included in the
distribution.

Automake doesn't support installing directories, so we use the
`install-data-hook` target to copy the gems directory to the right place
(prefix/share/gitsh/gems).