Fix vendorize process.
The improvements in 6f72fdd introduced a bug: If `vendor/gems` wasn't
present when `./configure` was run, then `$vendorfile` was empty, and so
`make dist` would vendorize and distribute the gems, but `make install`
wouldn't install them.

This change makes sure the vendorized gems are present in the configure
script, and will vendorize them if needed. The vendorized gems are included
in the distribution, so when installing from the tarball this is a no-op.