Add color placeholders (%c & %w) to prompt format
* %c begins coloring based on status
* %w ends coloring

This also inverts the way color support was handled: Previously we added
colors if the terminal supported them, now we add them by default and remove
them if the terminal lacks support. This is more flexible, because we can
add static color codes to the `gitsh.prompt` setting and the default prompt
without worrying about compatibility issues.