Remove hard coded colour from the default prompt.
Users can't add their own ANSI escape sequences to their `gitsh.prompt`
setting because of restrictions on escape sequences in `.gitconfig` files.
It's not great for the default prompt to include a feature the user can't
use in their custom prompt, so we'll remove it until we have support for
custom colours.