Fix homebrew install for Ruby < 2.0
Homebrew was detecting the system Ruby version correctly and adding a
dependency on Homebrew Ruby if it was < 2.0, but the configure script was
still heavily biased towards picking system Ruby.

This explicitly passed the Ruby path from the Homebrew formula to the
configure script to ensure the Ruby version chosen by Homebrew is used.