Mention gh, which is faster than hub
Showing `hub` in the example will cause people to try it with `hub`,
which will lead to an unpleasant first experience due to `hub`'s speed.
By showing `gh` in the example, we hint at something better and help
popularize the faster alternative.

We see this in FactoryGirl, where people complain that FG is slow merely
because they write to the database often.