Commit the generated INSTALL documentation
The `INSTALL` file is generated from `INSTALL.in`; however, since it is
the installation guide, it is often read before the program is even
downloaded. For example, on GitHub we link to `INSTALL.in` from the
`README.md`. Rather than recommend that people read a template of a
documentation, we can instead simply add the generated documentation to
the repo.