Consider unset variables to be an error.
If a command uses a variable that has not previously been set, the command
will not be executed, and an error message will be printed identifying the
unset variable.

Introduces a Gitsh::Error class and Gitsh::UnsetVariableError subclass to
represent errors that are fatal to the current command but not to the whole
gitsh session.