Lazily evaluate variables in arguments.
Instead of replacing variables with strings in the Gitsh::Transformer, only
evaluate their final value when they are needed by one of the
Gitsh::Commands::* objects.

This makes it possible to run commands like:

    :set foo 1 && :echo $foo

Introduces a Gitsh::ArgumentList to represent the arguments passed to a
command, various Gitsh::Arguments::* classes to represent the individual
arguments, and a Gitsh::ArgumentBuilder to construct arguments which can
have a complex structure consisting of literal parts and variable parts.