Update INSTALL in the configure script.
The way we were building the INSTALL instructions was causing `make
distcheck` to fail: `make clean` wasn't cleaning up the INSTALL file, which
was what we wanted for building in the source directory, but not when
building from another directory.

Letting autoconf handle the INSTALL file fixes this problem, and simplifies
the Makefile in the process.