Add a :source internal command.
`:source path` will run any Git and gitsh commands in the file at `path`. It
can modify the environment of the gitsh session that calls it, much like the
`source` command in Unix shells.

The implementation re-uses the ScriptRunner, which therefore needed to be
modified to raise on errors instead of exiting.

Paves the way for a future ~/.gitshrc file.