Reliably update terminal size with stty(1).
When gitsh receives a `SIGWINCH`, it informs Readline of the new terminal
size.

Previously, the new terminal size was determined using the `tput cols` and
`tput lines` commands. However, when Readline is sent the terminal size it
sets `COLUMNS` and `LINES` environment variables, and if these environment
variables are set `tput` will read them instead of actually checking the
terminal size. This means that we were never actually passing the new
terminal size to Readline, we were just passing whatever size we had when
the environment variables were set, over and over again.

Determining the terminal size by trying each of the following in turn seems
to work better:

1. `stty size` – this might fail, because some implementations of `stty`
   don't support `size`, e.g. on Solaris.
2. `env COLUMNS='' tput cols` and `env LINES='' tput lines` – this might
   fail, because some implementations of `tput` don't support `cols` or
   `lines` unless the environment variables are present.
3. `tput cols` and `tput lines` – this is what we have now.
4. Ignore the `SIGWINCH` entirely.

This also involved switching from using `Open3.capture3` to `IO.popen`,
because the `stty` command only works if its input stream is a TTY, and
`Open3.capture3` passes a pipe instead of `$stdin`. Unfortunately, this
makes the testing more complex.

Fixes #144. Probably. I can no longer produce the reproducible resize
error [1] after this commit, so it's promising.

[1]: https://github.com/thoughtbot/gitsh/issues/144#issuecomment-69459773