Ignore sequential duplicate history entries
Previously when the same command had been issued several times in a row,
users would have to press the up arrow an equal number of times to get
past it in the history. Some users found this tiresome.

This commit changes the Readline filter to filter out sequential
duplicate history entires in addition to blank ones.

This fixes #68.