WIP: Completer accepts quoted spaces in user input
The Readline side of this is done: the new Gitsh::QuoteDetector takes care
of letting Readline know if the word boundary character it's considering is
escaped.

The Gitsh::Completer side of this is not done: when it accepts a string with
escaped characters, it doesn't match any of the options.