Use completion_append_character
Before: we'd set completion_append_character to nil, and then append a space
to any completion option that wasn't a directory path (e.g., completions
might include "file.txt " and "dir/"). This meant we needed to manually
remove the space from the end of the options if we were completing a quoted
argument, and we needed to avoid escaping spaces at the end of a completion
when completing an unquoted argument.

After: we set completion_append_character dynamically based on the
completion options. If we have a single option that's a directory path, we
set it to nil. If we have multiple options, or a single option that's not a
directly path, we set it to space.