Show home directory as a tilde in the prompt
When gitsh is launched from the user's home directory, the basename of
the current working directory shown by the `%D` format string is
typically the same as the user's username, which is a potential source
of confusion as to what this prompt section shows. With this commit, the
home directory name is replaced by a tilde in the `%d` and `%D` format
strings, as is conventional for Unix shells.

Without this commit:

    $ cd ~
    $ pwd
    /home/srs
    $ gitsh
    gitsh 0.10
    Type :exit to exit
    srs uninitialized!!

With this commit:

    $ cd ~
    $ pwd
    /home/srs
    $ gitsh
    gitsh 0.10
    Type :exit to exit
    ~ uninitialized!!