Use a non-recursive Makefile.
Recursive Makefiles are less ideal [1] than iterative ones, for quite a few
reasons. But the main driver for this move is as a simplification so that we
can concentrate on the ext directory separately from autotools.

This commit does not fix the distcheck target, but dist and check each work.
Switching to a non-recursive Makefile has changed the location where some
files are installed. Rather than fighting with Automake to change their
install locations, this commit also updates the program to look for them in
their new homes.

$prefix/share/gitsh/gems -> $prefix/share/gitsh/vendor/gems
$prefix/share/gitsh/ruby/gitsh -> $prefix/share/gitsh/ruby/lib/gitsh

[1]: http://aegis.sourceforge.net/auug97.pdf