Honor DESTDIR during make install
Between version 0.10 and 0.11 `make` stopped honoring DESTDIR during the
install phase. This might not be relevant when installing from source
with the customary `sudo make install`, but becomes relevant when the
command is issued in package creation for distros. In such
circumstances, we don't want to install files directly in /usr/share but in
a custom directory that will be used to create the final package.