Refactor from runners to input strategies.
This commit replaces the `Gitsh::InteractiveRunner` and
`Gitsh::ScriptRunner` classes with input strategies
(`Gitsh::InputStrategies::Interactive` and `Gitsh::InputStrategies::File`).
Previously each runner class's `#run` method would loop over available input
and pass each command to the interpreter. Now the interpreter controls the
loop and asks the strategy for commands.

The main motivation is to support multi-line constructs. We would've needed
to add very similar code for handling incomplete input to both the runner
classes. Now that the interpreter is responsible for controlling to overall
flow of the session, we'll be able to handle incomplete input in one place.

A nice side-effect is that the `Gitsh::CLI` code is cleaner, and now uses
conditionals to create different strategies instead of using conditionals to
fork the control flow.

This commit also introduces some new `Gitsh::*Runner` classes, which take
over responsibility for the places where we still need to execute some gitsh
code outside of the main run loop:

* The old `Gitsh::ScriptRunner` class was used to load the `.gitshrc` file
  at the start of an interactive session, and by the internal `:source`
  command.  This has been replaced with a `Gitsh::FileRunner` that handles
  setting up a `Gitsh::InputStrategies::File` and a `Gitsh::Interpreter` for
  a given file.

* Similarly, the old `Gitsh::Interpreter#execute` method was used by
  `Gitsh::Arguments::Subshell` to run the contents of a subshell. This has
  been replaced with a `Gitsh::StringRunner` and associated
  `Gitsh::InputStrategies::String`.