Non-zero exit status for bad scripts.
When gitsh is invoked in a non-interactive manner, a parse error should
result in a non-zero exit status.

To make sure this happens, parse error handling has been delegated to the
input strategies by the interpreter. In an interactive session, we output an
error message and continue the session. In a non-interactive session, we
raise a Gitsh::ParseError, which will bubble up to the gitsh binary and
terminate the session with an exit status of 1.

Since the input strategy for files is used in a few internal places (in the
interactive input strategy, for loading .gitshrc, and in the :source
command), those call sites have been modified to handle the possibility of a
Gitsh::ParseError.