Support explicit architectures for Readline.
On macOS systems, universal binaries contain multiple architectures, e.g.
the system Ruby is built for both i386 and x86_64. The Ruby `mkmf.rb` build
system will default to building universal C extensions when used with a
universal Ruby.

When installing gitsh via Homebrew, we build against the Homebrew version of
Readline. Until recently, this was a universal library (x86_64 & i386) but
is now x86_64 only [1].

It is possible to build against a universal Ruby and a non-universal
Readline, but only if the right flags are passed to the C extension's
`extconf.rb` script.

This commit:

- Adds support for a `$READLINE_ARCH` environment variable. When it is set,
  the `configure` script will pass the right `--with-arch-flags` argument to
  `extconf.rb`.

- Updates the Homebrew formula to set `$READLINE_ARCH` based on the host
  OS's preferred architecture.

[1]: https://github.com/Homebrew/homebrew-core/pull/10907