Improve variable lexing.
Prior to this commit, an incomplete variable reference (a line ending with
an unescaped `$` character, or unclosed `${`) would produce a lexing error.
After this commit it produces a parse error instead.

We want to start using the lexer for understanding tab completion context,
which means we need to lex incomplete input. Having all legal-but-incomplete
input cause parse errors instead of lexing errors will make this much
easier.

Note that it's still possible to produce lexing errors with invalid input,
but it's more reasonable for the tab completion system to just ignore that
and not provide any completions.

Previously the token `MISSING` was used to indicate that a line was
incomplete, but could potentially be made valid with another line of input.
This commit renames that token to `INCOMPLETE`, and re-purposes `MISSING` to
mean that something was missing from the end of the line that cannot be
supplied by an additional line of input.