DSL to define tab completion automaton state graph
This commit introduces a domain specific language (DSL) to define the state
graph for the tab completion automaton. It also replaces the hard-coded
state graph with an equivalent graph loaded from a configuration file.

The details of the DSL are given in the gitsh_completions(5) manual page
which is added in this commit.

The DSL will need to be expanded somewhat to do everything we want the tab
completion system to be able to do, but this is a good step in that
direction.