Filter out duplicate completions.
Our ability to detect incomplete paths is based on there being a single
result (e.g. if you complete "s" to "src/" we don't want to append a space
as we normally would when there's a single completion).

Now that the completion system uses an NFA, it's possible to get duplicate
values (e.g. a command like `add $path* --? $path+` would invoke the `$path`
matcher twice in some situations).

This commit makes sure we filter out the duplicates so that the incomplete
path detection continues to work.

(While we're modifying the only line in this method, also move it to get
dependency order between #matches and #completions)