Add option support to the tab completion DSL.
This commit expands the tab completion DSL with the ability to define
options like `-f` or `--force`.

Rules can contain an `$opt` variable, which indicates where options may
appear. This will match on anything beginning with `-` or `--`, and
complete to any known options.

The known options are listed as indented lines after the first line of the
rule, with any arguments the options take specified in the same way as
the command's arguments would be.

For example:

```
commit $opt* --? $path+
  --all
  --fixup $revision
  --file $path
  --untracked-files (no|normal|all)?
```