Configure completions for common commands.
This tab completion configuration file includes:

- All Git commands without arguments.
- Popular Git commands, with their options and arguments.
- Internal gitsh commands, with their options and arguments.

There are some argument types that we can't currently complete, and others
that are unlikely to ever be practical to complete. To represent those this
commit also introduces a `$anything` variable to the tab completion DSL. The
variable's corresponding `AnythingMatcher` class will match any input, but
produce no completions.

The options and arguments supported in the file are based on the manual
pages from Git 2.13.3. Other versions of Git support slightly different
options and arguments, which means we might complete things that don't work.