Modernise Git command list for tab completion.
This list was originally built by parsing the output of `git help -a`, which
was based on the approach taken by the tab completion scripts distributed
with Git for various general-purpose shells.

In Git version 2.20, the output of `git help -a` changed in ways that
weren't compatible with our existing parsing code. Passing `--no-verbose`
continues to provide the old output format. Meanwhile, a new `--list-cmds`
argument was added, which outputs commands as a simple list, but is also
described in the git(1) manual page as being experimental and subject to
change.

To make sure we continue supporting a range of Git versions, and we won't
have major problems if the `--list-cmds` argument is removed, this commit
introduces the following strategy for getting a full list of Git commands:

1) Try `git --list-cmds`
2) Try `git help -a --no-verbose`
3) Try `git help -a`
4) Fail with an empty array