Fix tab completion with no input.
When there is no input, hitting tab twice should complete a list of all
available commands.

There was a bug in the Gitsh::TabCompletion::Context#completing_variable?
method which caused a crash in this case, because it would try to read the
type of the last token, which was `nil`.