Change process name by embedding Ruby in C
Fixes #41.

This resolves the process name issue by running a Ruby interpreter
in-process instead of via `ruby(1)`. The implementation is inspired by the
article [Running Ruby in C][1] and [Ruby's own `main` function][2].  Instead
of invoking `ruby(1)` as a subprocess, we use `ruby.h` and `libruby` to
perform the same function directly.

[1]: https://silverhammermba.github.io/emberb/embed/
[2]: https://github.com/ruby/ruby/blob/bdbc8a8f1214b601ca409f66f2fe00ac06f2e99a/main.c

In order to link against the correct version of Ruby, this commit also
introduces a `configure.rb` script which extracts information from
`RbConfig::MAKEFILE_CONFIG` so that it can be used in `configure.ac`.

Co-authored-by: Mike Burns <mburns@thoughtbot.com>
Co-authored-by: George Brocklehurst <george@georgebrock.com>