Reset $SAFE after line editor specs.
In Ruby 2.6, $SAFE changes from being a thread-local variable to a
process-local variable. This was causing the Gitsh::LineEditor specs to fail
under Ruby 2.6, because they set $SAFE to 1 in a thread, and never reset it
to 0, so the spec would fail at the next unsafe operation.

For more information, see: https://bugs.ruby-lang.org/issues/14250