Set rl_readline_name to "gitsh"
This change allows users to set conditional rules in their `.inputrc` that
only apply to gitsh.

Testing this relies on explicitly reloading the `.inputrc` via the `\cx\cr`
keyboard command. This is somewhat subtle, because running a test that does
not do this in isolation will load the `.inputrc` as defined in the test,
but running it as part of a larger test suite will inherit the pre-existing
`.inputrc`. It also depends on the system where the tests are run not having
any `.inputrc` rules that would disable the `\cx\cr` reload behaviour.

See: http://web.mit.edu/gnu/doc/html/rlman_1.html#SEC10