Expand aliases during tab completion.
When a user attempts to tab complete the arguments of an alias, they would
previously need a custom tab completion rule. This commit expands the alias
behind the scenes, so that the tab completion system is using all available
context, e.g. with `alias.a = add` the tab completion system will treat `a
foo<TAB>` as if it were `add foo<TAB>`, allowing the standard tab completion
rules for `add` to be used without any additional configuration.

The downside to this approach is that it makes it impossible to specify a
custom tab-completion rule for an alias. The upside is that that shouldn't
be necessary.