Tab completion fallback rules.
This commit adds fallback rules to the default tab completion configuration,
which will only be used when no other rule applies. These provide sensible
defaults for things that aren't covered by a specific rule, e.g. non-Git
commands like `!rm`, or custom Git commands that have not been configured.

THE PROBLEM
===========

Providing generic fallbacks wasn't possible without adding some new features
to the tab completion DSL.

To illustrate why the existing DSL was insufficient, consider the following
rules:

    add $path+

    $command ($path|$revision|$remote)+

A user trying to tab complete arguments to git-add(1) will match both rules:
the input will match the `add` rule, and the user will correctly be offered
paths to complete, but it will also match the `$command` rule, and so they
will also be offered revisions and remotes.

The generic rule provides us with sensible tab completion defaults in
situations where we don't have a specific rule, but by matching widely it
also undermines the usefulness of a context-aware system.

THE SOLUTION
============

This commit solves the problem by introducing a `fallback::` modifier which
indicates a variable should only be used when there's no other applicable
rule. In terms of the finite state automaton that's built from the DSL, a
variable with a `fallback::` modifier represents a state transition that
will be followed if and only if there's no other appropriate transition from
the current state.

We can use this modifier to fix the problems with the example above:

    add $path+

    fallback::$command ($path|$revision|$remote)+

Now we get the best of both worlds: tab completing arguments for git-add(1)
will match the specific rule and only be offered paths, but any other
command will use the fallback rule and be given a generic set of completions
including paths, revisions, and remotes.