Useful error messages for bad completion vars.
The tab completion configuration language supports variables which represent
matchers for particular types of arguments, e.g. `$branch` or `$path`.

This commit introduces error handling when a tab completion configuration
file contains a variable that isn't supported by gitsh. The error message
will include the name of the variable, the path to the file where the error
occurred, and the exact position in the file.