Prevent invalid option definitions.
Before this commit, the following tab completion rule was considered valid,
but would cause an infinite loop when used:

    add $opt
      --example $opt

This was also considered valid, but doesn't really make sense:

    add $opt
      --example --nested

This commit resolved the situation by modifying the parser to not accept the
`$opt` token or another option token (e.g. `--nested`) in the definition of
an option.

Unfortunately this complicates the parser somewhat, because the tokens
permitted in a rule and in an option definition have diverged.