Fix subshell internal command output.
When using the `:echo` command in a subshell, e.g. `:echo foo $(:echo bar)`,
the expected output is `foo bar`, but prior to this commit the actual output
was `bar\nfoo`.

This commit resolves this issue by adding `#puts` and `#print` methods to
the `Gitsh::CapturingEnvironment` class to make sure the output goes to the
right place.