Fix missing space after subshells.
A subshell in an argument list would previously consume its trailing
whitespace and end up concatenated to the next argument, for example:

    :echo 1 $(:echo 2) 3

This should output "1 2 3", but prior to this commit it would output "1 23".

This fix is a change to the lexer: the regular expression representing the
RIGHT_PAREN and SUBSHELL_END tokens has been changed to not include any
trailing space. This allows the space to be captured as a separate SPACE
token, which the parser will then correctly interpret as separating a
subshell from the next argument in the list.

The downside of this approach is that the parser now has to handle an
optional SPACE token after any SUBSHELL_END or RIGHT_PAREN token. Previously
there was a cleaner separation of space being mostly handled in the lexer.