Refactor: Argument values are always arrays.
This commit updates all of the various `Gitsh::Arguments::*` classes'
`#value` methods to produce Arrays instead of single values. The
`Gitsh::ArgumentList` class takes on the responsibility of flattening the
results.

This change paves the way for various kinds of globbing and argument
expansion which could allow a single argument to produce multiple values.
For example, in GNU Bash the argument `h{i,o}p` would be expanded to the two
arguments `hip hop`.