Fix flaky test.
Setting `ENV['INPUTRC']` to `"/dev/null"` and back to `nil` in the
`Gitsh::LineEditor` spec was causing the `.inputrc` integration spec to fail
if it happened to run afterwards.

This commit avoids the problem by leaving the `INPUTRC` variable untouched.
The code to change it was inherited from the Ruby `Readline` tests, and it
seems unlikely we'll end up with a contributor with a custom `INPUTRC` path
that we need to work around.