Refactor: Decide command class at execution time.
This commit replaces the `Gitsh::Commands::Factory` class with a
`Gitsh::Commands::LazyCommand` class. The logic to construct a command is
exactly the same, the difference is that the `LazyCommand` instance runs
that logic immediately before the command is executed, whereas the `Factory`
instance ran it at parse time.

The intention behind this refactor is to pave the way for commands that are
dynamically built at execution time. Future commits will build on this to
add the ability to specify the name of a command in a variable, and
eventually these changes will make it easier to implement support for
globbing patterns.