Argument zero ain't nothin' special.
This commit updates the Parser and LazyCommand to treat the name of any
command like any other argument.

Prior to this commit argument zero had to be a constant string. After this
commit it's possible to use a variable, subshell, or any other dynamic value
we might introduce as argument zero.

To give a concrete example, we can now do this:

    :set command :set
    $command greeting Hello
    :echo "$greeting, world"

This is more in line with how general purpose Unix shells operate. It also
helps pave the way for supporting more types of dynamic arguments, like glob
patterns.