Run the compiled binary on CI
Now that we embed Ruby in the gitsh binary, it's worth running the built
binary on CI so that we can catch any regressions that occur if the C API
changes in future versions.

Setting the LD_LIBRARY_PATH is needed because rvm rubies are installed to a
non-standard location, and either Travis or rvm itself doesn't handle
setting this for us.

Co-authored-by: Mike Burns <mburns@thoughtbot.com>