Fix tab completing paths with no prefix.
When tab completing a path argument without entering the beginning of the
path we expect the completion options to be the files in the current
directory.

Prior to this commit the only option would be `/`, until the user entered at
least the first letter of the path to be completed.