Support installing gitsh from HEAD via Homebrew
Supersedes https://github.com/thoughtbot/homebrew-formulae/pull/60.

Adds a `head` configuration block to the Homebrew formula to support
top-of-tree builds.

Additionally adds a VERSION_SUFFIX variable to configure.ac, which is
appended to PACKAGE_VERSION on `./configure`. This is used to produce a
SemVer-compatible version number such as 0.14-head+5d9484d, in order to
differentiate pre-release builds from official releases.